+++ firmware/lib/integration/tests/protocol/fast_bulk_read.rs
@@ -14,7 +14,7 @@ use osc_integration::sim::{HOST_INTER_BYTE_TIMEOUT, SimTime};
 use rstest::rstest;
 use rstest_reuse::apply;
 
-const CONFIG_REGION_END_ADDR: u16 = CONFIG_REGION_SIZE as u16;
+const CONFIG_REGION_END_ADDR: u16 = CONFIG_REGION_SIZE;
 const OVER_MAX_CONTROL_RW: u16 = MAX_CONTROL_RW as u16 + 1;
 
 fn entry(id: u8, address: u16, length: u16) -> BulkReadEntry {
